+++ docs/source/conf.py
@@ -39,11 +39,7 @@ sys.path.insert(0, os.path.abspath('../src/'))
 # ones.
 extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.mathjax',
-              'sphinx.ext.viewcode']#,
-#              'sphinx.ext.autosummary']#,
-#              'm2r']
-
-#autosummary_generate = True
+              'sphinx.ext.viewcode']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -145,7 +141,7 @@ latex_elements = {
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
     (master_doc, 'data_tools.tex', u'data\\_tools Documentation',
-     u'Nicolàs Palacio', 'manual'),
+     u'Nicolàs Palacio-Escat', 'manual'),
 ]