+++ src/data_tools/__info__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+__version__ = u'0.0.6'
+__author__ = u'Nicolàs Palacio-Escat'